x86/bugs: Add /sys/devices/systen/cpu/vulnerabilities/CVE_list
Add a sysfs file listing the CVEs that are known or believed to be
addressed by this kernel.

The existing content of the cpu/vulnerabilities/ directory shows a
list of the mitigations that are implemented by this kernel. However,
it does not give an easy way to correlate these entries to the CVE
database.

This commit adds a list of the CVEs that are known to be addressed by
the kernel.

Signed-off-by: Christophe de Dinechin <dinechin@redhat.com>